Changes and additions to improve aiming accuracy (~50% accurate currently)

+++ Curve.cpp
@@ -2,44 +2,44 @@
 
 const unsigned int NUM_SMOOTHING_ITERATIONS = 10;
 
-vec2<int> Curve::linear(std::vector<vec2<int>> controlPoints, double t)
+vec2<double> Curve::linear(std::vector<vec2<double>> controlPoints, double t)
 {
-	vec2<int> x = controlPoints[0];
-	vec2<int> y = controlPoints[1];
+	vec2<double> x = controlPoints[0];
+	vec2<double> y = controlPoints[1];
 
 	if (x == y)
 		return x;
 
-	int xDir = y.getX() - x.getX();
-	int yDir = y.getY() - x.getY();
+	double xDir = y.getX() - x.getX();
+	double yDir = y.getY() - x.getY();
 
 	// unsigned magnitude = (y - x).magnitude();
 	
-	return vec2<int>(x.getX() + static_cast<int>(xDir * t), x.getY() + static_cast<int>(yDir * t));
+	return vec2<double>(x.getX() + static_cast<double>(xDir * t), x.getY() + static_cast<double>(yDir * t));
 }
 
-vec2<int> Curve::bezier2(std::vector<vec2<int>> controlPoints, double t)
+vec2<double> Curve::bezier2(std::vector<vec2<double>> controlPoints, double t)
 {
-	vec2<int> x = controlPoints[0];
-	vec2<int> y = controlPoints[1];
-	vec2<int> z = controlPoints[2];
+	vec2<double> x = controlPoints[0];
+	vec2<double> y = controlPoints[1];
+	vec2<double> z = controlPoints[2];
 
 	double t2 = t * t;
 	double mt = 1 - t;
 	double mt2 = mt * mt;
 
-	int a = static_cast<int>(mt2 * x.getX() + 2 * mt * t * y.getX() + t2 * z.getX());
-	int b = static_cast<int>(mt2 * x.getY() + 2 * mt * t * y.getY() + t2 * z.getY());
+	double a = (mt2 * x.getX() + 2 * mt * t * y.getX() + t2 * z.getX());
+	double b = (mt2 * x.getY() + 2 * mt * t * y.getY() + t2 * z.getY());
 
-	return vec2<int>{ a, b };
+	return vec2<double>{ a, b };
 }
 
-vec2<int> Curve::bezier3(std::vector<vec2<int>> controlPoints, double t)
+vec2<double> Curve::bezier3(std::vector<vec2<double>> controlPoints, double t)
 {
-	vec2<int> w = controlPoints[0];
-	vec2<int> x = controlPoints[1];
-	vec2<int> y = controlPoints[2];
-	vec2<int> z = controlPoints[3];
+	vec2<double> w = controlPoints[0];
+	vec2<double> x = controlPoints[1];
+	vec2<double> y = controlPoints[2];
+	vec2<double> z = controlPoints[3];
 
 	double t2 = t * t;
 	double t3 = t2 * t;
@@ -47,13 +47,13 @@ vec2<int> Curve::bezier3(std::vector<vec2<int>> controlPoints, double t)
 	double mt2 = mt * mt;
 	double mt3 = mt2 * mt;
 
-	int a = static_cast<int>(mt3 * w.getX() + 3 * mt2 * t * x.getX() + 3 * mt2 * t2 * y.getX() + t3 * z.getX());
-	int b = static_cast<int>(mt3 * w.getY() + 3 * mt2 * t * x.getY() + 3 * mt2 * t2 * y.getY() + t3 * z.getY());
+	double a = (mt3 * w.getX() + 3 * mt2 * t * x.getX() + 3 * mt2 * t2 * y.getX() + t3 * z.getX());
+	double b = (mt3 * w.getY() + 3 * mt2 * t * x.getY() + 3 * mt2 * t2 * y.getY() + t3 * z.getY());
 
-	return vec2<int>{ a, b };
+	return vec2<double>{ a, b };
 }
 
-void Curve::bezierPath(std::vector<vec2<int>> &sliderPoints, std::vector<vec2<int>> controlPoints)
+void Curve::bezierPath(std::vector<vec2<double>> &sliderPoints, std::vector<vec2<double>> controlPoints)
 {	
 	// osu! sliders made up of more than one line/curve are denoted by repeated sequential points
 	// where one curve ends and the next one starts. To coalesce curves simply calculate 
@@ -71,19 +71,19 @@ void Curve::bezierPath(std::vector<vec2<int>> &sliderPoints, std::vector<vec2<in
 		// Peek to see if start of next curve
 		else if ((end != controlPoints.size() - 1 && controlPoints[end] == controlPoints[end + 1]) || end == controlPoints.size() - 1) {
 			if (end - start < 2) { // 2 total points -> linear
-				auto points = std::vector<vec2<int>>{ controlPoints[start], controlPoints[end] };
+				auto points = std::vector<vec2<double>>{ controlPoints[start], controlPoints[end] };
 				for (unsigned i = 1; i <= NUM_SMOOTHING_ITERATIONS; i++) {
 					sliderPoints.push_back(linear(points, 1.0 / NUM_SMOOTHING_ITERATIONS * i));
 				}
 			}
 			else if (end - start < 3) { // 3 total points -> bezier2
-				auto points = std::vector<vec2<int>>{ controlPoints[start], controlPoints[start + 1], controlPoints[end] };
+				auto points = std::vector<vec2<double>>{ controlPoints[start], controlPoints[start + 1], controlPoints[end] };
 				for (unsigned i = 1; i <= NUM_SMOOTHING_ITERATIONS; i++) {
 					sliderPoints.push_back(bezier2(points, 1.0 / NUM_SMOOTHING_ITERATIONS * i));
 				}
 			}
 			else { // 4 total points -> bezier3
-				auto points = std::vector<vec2<int>>{ controlPoints[start], controlPoints[start + 1], controlPoints[start+2], controlPoints[end] };
+				auto points = std::vector<vec2<double>>{ controlPoints[start], controlPoints[start + 1], controlPoints[start+2], controlPoints[end] };
 				for (unsigned i = 1; i <= NUM_SMOOTHING_ITERATIONS; i++) {
 					sliderPoints.push_back(bezier3(points, 1.0 / NUM_SMOOTHING_ITERATIONS * i));
 				}
@@ -93,9 +93,9 @@ void Curve::bezierPath(std::vector<vec2<int>> &sliderPoints, std::vector<vec2<in
 	}
 }
 
-std::vector<vec2<int>> Curve::generateSliderPoints(std::vector<vec2<int>> controlPoints, unsigned numPoints)
+std::vector<vec2<double>> Curve::generateSliderPoints(std::vector<vec2<double>> controlPoints, unsigned numPoints)
 {
-	std::vector<vec2<int>> sliderPoints;
+	std::vector<vec2<double>> sliderPoints;
 	sliderPoints.push_back(controlPoints[0]);
 
 	switch (numPoints) {
