+++ HitCircle.h
@@ -5,7 +5,7 @@ class HitCircle :
 	public HitObject
 {
 public:
-	HitCircle(int x, int y, unsigned startTime, unsigned endTime, unsigned type);
+	HitCircle(double x, double y, unsigned startTime, unsigned endTime, unsigned type);
 	~HitCircle();
 
 	virtual void mouseMovement(HANDLE osuProcess, DWORD timeAddress);
