+++ Spinner.h
@@ -5,7 +5,7 @@ class Spinner :
 	public HitObject
 {
 public:
-	Spinner(int x, int y, unsigned startTime, unsigned endTime, unsigned type);
+	Spinner(double x, double y, unsigned startTime, unsigned endTime, unsigned type);
 	~Spinner();
 
 	virtual void mouseMovement(HANDLE osuProcess, DWORD timeAddress);
